ルートモジュールでのモジュール分割対応
