+++ backend/user/build.gradle.kts
@@ -28,34 +28,16 @@ repositories {
 }
 
 dependencies {
-    implementation("io.ktor:ktor-server-core-jvm")
-    implementation("io.ktor:ktor-server-auth-jvm")
-    implementation("io.ktor:ktor-server-auth-jwt-jvm")
-    implementation("io.ktor:ktor-server-auto-head-response-jvm")
-    implementation("io.ktor:ktor-server-host-common-jvm")
-    implementation("io.ktor:ktor-server-status-pages-jvm")
-    implementation("io.ktor:ktor-server-compression-jvm")
-    implementation("io.ktor:ktor-server-cors-jvm")
-    implementation("io.ktor:ktor-server-call-logging-jvm")
-    implementation("io.ktor:ktor-server-content-negotiation-jvm")
-    implementation("io.ktor:ktor-serialization-jackson")
-    implementation("org.jetbrains.exposed:exposed-core:$exposed_version")
-    implementation("org.jetbrains.exposed:exposed-jdbc:$exposed_version")
-    implementation("org.jetbrains.exposed:exposed-kotlin-datetime:$exposed_version")
-    implementation("com.mysql:mysql-connector-j:8.4.0")
-    implementation("io.ktor:ktor-server-netty-jvm")
-    implementation("ch.qos.logback:logback-classic:$logback_version")
-    implementation(platform("io.insert-koin:koin-bom:3.5.6"))
-    implementation("io.insert-koin:koin-core")
-    implementation("io.insert-koin:koin-ktor")
-    implementation("io.insert-koin:koin-logger-slf4j")
-    implementation("io.insert-koin:koin-annotations:1.3.1")
-    implementation("at.favre.lib:bcrypt:0.10.2")
-    implementation("io.arrow-kt:arrow-core:1.2.4")
-    ksp("io.insert-koin:koin-ksp-compiler:1.3.1")
-    testImplementation("io.ktor:ktor-server-tests-jvm")
-    testImplementation("org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version")
-    testImplementation("io.insert-koin:koin-test")
+    implementation(libs.bundles.ktor)
+    implementation(libs.logback)
+    implementation(libs.bundles.koin)
+    implementation(libs.koin.ktor)
+    implementation(libs.arrow)
+    implementation(project(":repository"))
+    testImplementation(libs.bundles.kotest)
+    testImplementation(libs.test.kotlin)
+    testImplementation(libs.test.koin)
+    testImplementation(libs.test.ktor)
 }
 
 ksp {
