+++ backend/user/entity/src/main/kotlin/myapp/entity/User.kt
@@ -5,13 +5,3 @@ data class User(
     val name: String,
     val password: String,
 )
-
-data class UserLogin(
-    val name: String,
-    val password: String,
-)
-
-data class UserPublic(
-    val id: Int,
-    val name: String,
-)
