+++ backend/user/repository/src/main/kotlin/myapp/repository/Repository.kt
@@ -7,8 +7,20 @@ import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransacti
 import org.koin.dsl.bind
 import org.koin.dsl.module
 
-val repositoryModuke = module {
+val repositoryModule = module {
+    single { databaseConfig() }
     single { UserRepositoryImpl(get()) }.bind(UserRepository::class)
 }
+
+fun databaseConfig(
+    dbUrl: String = System.getenv("DB_URL") ?: "jdbc:mysql://localhost:3306/training",
+    user: String = System.getenv("DB_USER") ?: "root",
+    password: String = System.getenv("DB_PASSWORD") ?: "",
+) = Database.connect(
+    url = dbUrl,
+    user = user,
+    password = password,
+    driver = "com.mysql.cj.jdbc.Driver",
+)
 
 internal suspend fun <T> Database.query(block: suspend () -> T): T = newSuspendedTransaction(Dispatchers.IO, this) { block() }
