+++ backend/user/src/main/kotlin/myapp/model/User.kt
@@ -1,11 +1,5 @@
 package myapp.model
 
-data class User(
-    val id: Int,
-    val name: String,
-    val password: String,
-)
-
 data class UserLogin(
     val name: String,
     val password: String,
